+++ ethereum/contracts/CDPManagerHelper.ts
@@ -1,9 +1,9 @@
 import { DSProxy__factory, DssCdpManager__factory } from 'generated/types';
 
-import IlkType from './IlkType';
-import { INT_FORMAT, toBigNumber } from './math';
+import IlkType from '../IlkType';
+import { INT_FORMAT, toBigNumber } from '../math';
 
-import type EthereumProvider from './EthereumProvider';
+import type EthereumProvider from '../EthereumProvider';
 import type { CDP } from './GetCDPsHelper';
 import type { FixedNumber } from 'ethers';
 import type { DssCdpManager } from 'generated/types';
