+++ ethereum/contracts/GetCDPsHelper.ts
@@ -1,10 +1,10 @@
-import IlkType from 'contracts/IlkType';
+import IlkType from 'ethereum/IlkType';
 import { GetCdps__factory } from 'generated/types';
 
-import { INT_FORMAT, toFixedNumber } from './math';
+import { INT_FORMAT, toFixedNumber } from '../math';
 
+import type EthereumProvider from '../EthereumProvider';
 import type CDPManagerHelper from './CDPManagerHelper';
-import type EthereumProvider from './EthereumProvider';
 import type { FixedNumber } from 'ethers';
 import type { GetCdps, DSProxy } from 'generated/types';
