+++ ethereum/contracts/VatHelper.ts
@@ -1,9 +1,9 @@
 import { Vat__factory } from 'generated/types';
 
-import { toFixedNumber, UnitFormats } from './math';
+import { toFixedNumber, UnitFormats } from '../math';
 
-import type EthereumProvider from './EthereumProvider';
-import type IlkType from './IlkType';
+import type EthereumProvider from '../EthereumProvider';
+import type IlkType from '../IlkType';
 import type { FixedNumber } from 'ethers';
 import type { Vat } from 'generated/types';
