+++ ethereum/react/cards/BNText.tsx
