+++ pages/forms/BurnForm.tsx
@@ -1,11 +1,11 @@
 import { Button, Card, Grid, InputAdornment, TextField, CircularProgress } from '@mui/material';
 import { useCallback, useMemo, useState } from 'react';
 
-import { UnitFormats } from 'contracts/math';
+import { UnitFormats } from 'ethereum/math';
 
 import { cutDecimals, pickNumbers, toFixedNumberOrUndefined } from './stringNumber';
 
-import type { IlkInfo } from 'contracts/IlkRegistryHelper';
+import type { IlkInfo } from 'ethereum/contracts/IlkRegistryHelper';
 import type { FixedNumber } from 'ethers';
 import type { FC, ChangeEventHandler, MouseEventHandler, ReactNode } from 'react';
