+++ pages/ilks/[ilk]/index.page.tsx
@@ -4,19 +4,19 @@ import Link from 'next/link';
 import { useRouter } from 'next/router';
 import { useCallback, useMemo } from 'react';
 
-import IlkType from 'contracts/IlkType';
-import Vault from 'contracts/Vault';
-import { useChainLog } from 'pages/ethereum/ContractHooks';
+import IlkType from 'ethereum/IlkType';
+import Vault from 'ethereum/Vault';
+import { useChainLog } from 'ethereum/react/ContractHooks';
+import IlkStatusCard, { useIlkStatusCardProps } from 'ethereum/react/cards/IlkStatusCard';
+import WalletStatusCard from 'ethereum/react/cards/WalletStatusCard';
 import MintForm from 'pages/forms/MintForm';
-import IlkStatusCard, { useIlkStatusCardProps } from 'pages/ilks/[ilk]/IlkStatusCard';
 import { getStringQuery } from 'pages/query';
 import usePromiseFactory from 'pages/usePromiseFactory';
-import WalletStatusCard from 'pages/vaults/[id]/WalletStatusCard';
 
-import type ChainLogHelper from 'contracts/ChainLogHelper';
-import type EthereumProvider from 'contracts/EthereumProvider';
-import type { IlkInfo } from 'contracts/IlkRegistryHelper';
-import type { IlkStatus } from 'contracts/VatHelper';
+import type EthereumProvider from 'ethereum/EthereumProvider';
+import type ChainLogHelper from 'ethereum/contracts/ChainLogHelper';
+import type { IlkInfo } from 'ethereum/contracts/IlkRegistryHelper';
+import type { IlkStatus } from 'ethereum/contracts/VatHelper';
 import type { FixedNumber } from 'ethers';
 import type { NextPageWithEthereum } from 'next';
 import type { FC } from 'react';
